fix: renamed file names according to project style
